+++ tests/integ_tests/workflow_tests/continue_as_new.rs
@@ -1,7 +1,7 @@
 use std::time::Duration;
 use temporal_sdk_core::{
     protos::coresdk::workflow_commands::{ContinueAsNewWorkflowExecution, StartTimer},
-    test_workflow_driver::{WfContext, WfExitValue, WorkflowResult},
+    prototype_rust_sdk::{WfContext, WfExitValue, WorkflowResult},
 };
 use test_utils::CoreWfStarter;
