+++ tests/integ_tests/workflow_tests/timers.rs
@@ -3,7 +3,7 @@ use temporal_sdk_core::protos::coresdk::{
     workflow_commands::{CancelTimer, CompleteWorkflowExecution, StartTimer},
     workflow_completion::WfActivationCompletion,
 };
-use temporal_sdk_core::test_workflow_driver::{WfContext, WorkflowResult};
+use temporal_sdk_core::prototype_rust_sdk::{WfContext, WorkflowResult};
 use test_utils::{init_core_and_create_wf, CoreTestHelpers, CoreWfStarter};
 
 pub async fn timer_wf(mut command_sink: WfContext) -> WorkflowResult<()> {
